change amount type to Long on ExpenseModel

+++ app/src/main/java/com/example/denip/nasyiatulaisyiyahfinance/expense/ExpenseModel.kt
@@ -6,13 +6,13 @@ package com.example.denip.nasyiatulaisyiyahfinance.expense
 
 class ExpenseModel(val addedByTreasurer: String?,
                    val addedByTreasurerUid: String?,
-                   val amount: Int?,
+                   val amount: Long?,
                    val category: String?,
                    val dateCreated: String?,
                    val expenseId: String?,
                    val note: String?,
                    val addedByTreasurerInitial: String?,
                    val categoryId: String?) {
-    constructor() : this("","", 0, "", "", "", "", "", "")
+    constructor() : this("", "", 0, "", "", "", "", "", "")
     constructor(addedByTreasurer: String?) : this(addedByTreasurer, "", 0, "", "", "", "", "", "")
 }
